// Angular/Sessions Material/D04/d04/src/app/types.ts
export interface Product {
  id: string;
  name: string;
  category: string;
  description?: string;
  price: number | string;
  quantity: number | string;
  imageUrl?: string;
  isInCart: boolean;
  status?: status;
}

type status = 'active' | 'inactive';

export type error = {
  message: string;
  state: boolean;
};

export type tabName = 'all' | 'cart';